Create return image
Added yellow.svg to use a return to home image.
While updating the html to adjust return to home link,
I found that one of the html files did not </body> in
the proper position; thus, that was updated. Added
to the style sheet to adjust the yellow.svg file.